Change import statement to reflect the move of SolrQueryConverter from query package to solr package.
